DevEx: break caseDetailHelper up into smaller computeds
