+++ web-client/src/presenter/computeds/caseDetailHelper.js
@@ -13,7 +13,6 @@ export const caseDetailHelper = (get, applicationContext) => {
     .getUtilities()
     .isExternalUser(user.role);
   const userAssociatedWithCase = get(state.screenMetadata.isAssociated);
-  const modalState = get(state.modal);
   let showEditPetitionerInformation = false;
   const permissions = get(state.permissions);
   const showJudgesNotes = permissions.JUDGES_NOTES;
@@ -61,34 +60,6 @@ export const caseDetailHelper = (get, applicationContext) => {
     showEditPetitionerInformation = true;
   }
 
-  const practitionerMatchesFormatted =
-    modalState && modalState.practitionerMatches;
-  if (practitionerMatchesFormatted) {
-    practitionerMatchesFormatted.map(practitioner => {
-      if (practitioner.contact) {
-        practitioner.cityStateZip = `${practitioner.contact.city}, ${practitioner.contact.state} ${practitioner.contact.postalCode}`;
-      }
-      if (caseDetail.privatePractitioners) {
-        practitioner.isAlreadyInCase = caseDetail.privatePractitioners.find(
-          casePractitioner => casePractitioner.userId === practitioner.userId,
-        );
-      }
-    });
-  }
-  const respondentMatchesFormatted = modalState && modalState.respondentMatches;
-  if (respondentMatchesFormatted) {
-    respondentMatchesFormatted.map(respondent => {
-      if (respondent.contact) {
-        respondent.cityStateZip = `${respondent.contact.city}, ${respondent.contact.state} ${respondent.contact.postalCode}`;
-      }
-      if (caseDetail.irsPractitioners) {
-        respondent.isAlreadyInCase = caseDetail.irsPractitioners.find(
-          caseRespondent => caseRespondent.userId === respondent.userId,
-        );
-      }
-    });
-  }
-
   const hasConsolidatedCases = !isEmpty(caseDetail.consolidatedCases);
 
   const petitionDocument = applicationContext
@@ -100,16 +71,6 @@ export const caseDetailHelper = (get, applicationContext) => {
     caseDeadlines,
     documentDetailTab,
     hasConsolidatedCases,
-    practitionerMatchesFormatted,
-    practitionerSearchResultsCount:
-      modalState &&
-      modalState.practitionerMatches &&
-      modalState.practitionerMatches.length,
-    respondentMatchesFormatted,
-    respondentSearchResultsCount:
-      modalState &&
-      modalState.respondentMatches &&
-      modalState.respondentMatches.length,
     showAddCorrespondenceButton: permissions.CASE_CORRESPONDENCE,
     showCaseDeadlinesExternal,
     showCaseDeadlinesInternal,
