+++ web-client/src/presenter/computeds/caseDetailHelper.test.js
@@ -442,190 +442,6 @@ describe('case detail computed', () => {
     expect(result.showRespondentSection).toEqual(false);
   });
 
-  it('should format practitioner matches with cityStateZip string and isAlreadyInCase boolean', () => {
-    const user = {
-      role: ROLES.privatePractitioner,
-      userId: '123',
-    };
-    const result = runCompute(caseDetailHelper, {
-      state: {
-        ...getBaseState(user),
-        caseDetail: { documents: [], privatePractitioners: [{ userId: '2' }] },
-        form: {},
-        modal: {
-          practitionerMatches: [
-            {
-              contact: { city: 'Somewhere', postalCode: '12345', state: 'AL' },
-              name: '1',
-              userId: '1',
-            },
-            {
-              contact: {
-                city: 'Somewhere Else',
-                postalCode: '54321',
-                state: 'TX',
-              },
-              name: '2',
-              userId: '2',
-            },
-          ],
-        },
-      },
-    });
-    expect(result.practitionerMatchesFormatted).toMatchObject([
-      {
-        cityStateZip: 'Somewhere, AL 12345',
-        name: '1',
-      },
-      {
-        cityStateZip: 'Somewhere Else, TX 54321',
-        name: '2',
-      },
-    ]);
-    expect(result.practitionerMatchesFormatted[0].isAlreadyInCase).toBeFalsy();
-    expect(result.practitionerMatchesFormatted[1].isAlreadyInCase).toBeTruthy();
-  });
-
-  it('should set practitionerSearchResultsCount to the length of the state.modal.practitionerMatches', () => {
-    const user = {
-      role: ROLES.privatePractitioner,
-      userId: '123',
-    };
-    const result = runCompute(caseDetailHelper, {
-      state: {
-        ...getBaseState(user),
-        caseDetail: { documents: [] },
-        form: {},
-        modal: { practitionerMatches: [{ name: '1' }, { name: '2' }] },
-      },
-    });
-    expect(result.practitionerSearchResultsCount).toEqual(2);
-  });
-
-  it('should set practitionerSearchResultsCount to 0 if the state.modal.practitionerMatches is an empty array', () => {
-    const user = {
-      role: ROLES.privatePractitioner,
-      userId: '123',
-    };
-    const result = runCompute(caseDetailHelper, {
-      state: {
-        ...getBaseState(user),
-        caseDetail: { documents: [] },
-        form: {},
-        modal: { practitionerMatches: [] },
-      },
-    });
-    expect(result.practitionerSearchResultsCount).toEqual(0);
-  });
-
-  it('should not set practitionerSearchResultsCount if state.modal is an empty object', () => {
-    const user = {
-      role: ROLES.privatePractitioner,
-      userId: '123',
-    };
-    const result = runCompute(caseDetailHelper, {
-      state: {
-        ...getBaseState(user),
-        caseDetail: { documents: [] },
-        form: {},
-        modal: {},
-      },
-    });
-    expect(result.practitionerSearchResultsCount).toBeUndefined();
-  });
-
-  it('should format respondent matches with cityStateZip string and isAlreadyInCase boolean', () => {
-    const user = {
-      role: ROLES.privatePractitioner,
-      userId: '123',
-    };
-    const result = runCompute(caseDetailHelper, {
-      state: {
-        ...getBaseState(user),
-        caseDetail: { documents: [], irsPractitioners: [{ userId: '1' }] },
-        form: {},
-        modal: {
-          respondentMatches: [
-            {
-              contact: { city: 'Somewhere', postalCode: '12345', state: 'AL' },
-              name: '1',
-              userId: '1',
-            },
-            {
-              contact: {
-                city: 'Somewhere Else',
-                postalCode: '54321',
-                state: 'TX',
-              },
-              name: '2',
-              userId: '2',
-            },
-          ],
-        },
-      },
-    });
-    expect(result.respondentMatchesFormatted).toMatchObject([
-      {
-        cityStateZip: 'Somewhere, AL 12345',
-        name: '1',
-      },
-      {
-        cityStateZip: 'Somewhere Else, TX 54321',
-        name: '2',
-      },
-    ]);
-    expect(result.respondentMatchesFormatted[0].isAlreadyInCase).toBeTruthy();
-    expect(result.respondentMatchesFormatted[1].isAlreadyInCase).toBeFalsy();
-  });
-
-  it('should set respondentSearchResultsCount to the length of the state.modal.respondentMatches', () => {
-    const user = {
-      role: ROLES.privatePractitioner,
-      userId: '123',
-    };
-    const result = runCompute(caseDetailHelper, {
-      state: {
-        ...getBaseState(user),
-        caseDetail: { documents: [] },
-        form: {},
-        modal: { respondentMatches: [{ name: '1' }, { name: '2' }] },
-      },
-    });
-    expect(result.respondentSearchResultsCount).toEqual(2);
-  });
-
-  it('should set respondentSearchResultsCount to 0 if the state.modal.respondentMatches is an empty array', () => {
-    const user = {
-      role: ROLES.privatePractitioner,
-      userId: '123',
-    };
-    const result = runCompute(caseDetailHelper, {
-      state: {
-        ...getBaseState(user),
-        caseDetail: { documents: [] },
-        form: {},
-        modal: { respondentMatches: [] },
-      },
-    });
-    expect(result.respondentSearchResultsCount).toEqual(0);
-  });
-
-  it('should not set respondentSearchResultsCount if state.modal is an empty object', () => {
-    const user = {
-      role: ROLES.privatePractitioner,
-      userId: '123',
-    };
-    const result = runCompute(caseDetailHelper, {
-      state: {
-        ...getBaseState(user),
-        caseDetail: { documents: [] },
-        form: {},
-        modal: {},
-      },
-    });
-    expect(result.respondentSearchResultsCount).toBeUndefined();
-  });
-
   it('should show empty state for consolidated cases', () => {
     const user = {
       role: ROLES.privatePractitioner,
