+++ web-client/src/presenter/computeds/caseDetailPractitionerSearchHelper.js
@@ -1,0 +1,47 @@
+import { state } from 'cerebral';
+
+export const caseDetailPractitionerSearchHelper = get => {
+  const caseDetail = get(state.caseDetail);
+  const modalState = get(state.modal);
+
+  const practitionerMatchesFormatted =
+    modalState && modalState.practitionerMatches;
+  if (practitionerMatchesFormatted) {
+    practitionerMatchesFormatted.map(practitioner => {
+      if (practitioner.contact) {
+        practitioner.cityStateZip = `${practitioner.contact.city}, ${practitioner.contact.state} ${practitioner.contact.postalCode}`;
+      }
+      if (caseDetail.privatePractitioners) {
+        practitioner.isAlreadyInCase = caseDetail.privatePractitioners.find(
+          casePractitioner => casePractitioner.userId === practitioner.userId,
+        );
+      }
+    });
+  }
+  const respondentMatchesFormatted = modalState && modalState.respondentMatches;
+  if (respondentMatchesFormatted) {
+    respondentMatchesFormatted.map(respondent => {
+      if (respondent.contact) {
+        respondent.cityStateZip = `${respondent.contact.city}, ${respondent.contact.state} ${respondent.contact.postalCode}`;
+      }
+      if (caseDetail.irsPractitioners) {
+        respondent.isAlreadyInCase = caseDetail.irsPractitioners.find(
+          caseRespondent => caseRespondent.userId === respondent.userId,
+        );
+      }
+    });
+  }
+
+  return {
+    practitionerMatchesFormatted,
+    practitionerSearchResultsCount:
+      modalState &&
+      modalState.practitionerMatches &&
+      modalState.practitionerMatches.length,
+    respondentMatchesFormatted,
+    respondentSearchResultsCount:
+      modalState &&
+      modalState.respondentMatches &&
+      modalState.respondentMatches.length,
+  };
+};
